dynamic network switcher based on env
